Place assuan calls in a different thread
* src/client.cc (client_t::call_assuan): Use call_assuan_async. Void.
 (call_assuan_async): New. To be used in CreateThread.

--
Using a different thread makes sense here as starting the uiserver
can take a while and the return value was ignored anyway.

This also fixes interaction with Kleopatra from Qt5 as the
wId handling code Kleopatra uses to make the explorer the
parent Window utilizes SendMessage calls which resulted in
a deadlock.