Move the GpgEX icon to the first entry
* src/gpgex.cc (gpgex_t::QueryContextMenu): Move the icon to
the sign & encrypt / decrypt & verify action.

--
This makes more sense because that is by far the most
prominent option we need. Also it seems usual to have
the icon at the top and then actions below that related
to the icon. At least dropbox also does it this way
to avoid the uglyness that the icon is always shown
twice.